refactored AbstractContextConfigurationReader moved content of AbstractContextConfigurationReader into ContextConfigurationReader removed AbstractContextConfigurationReader
